Move unit tests in their own directory
Will avoid mixing up integration tests and unit tests
