+++ tests/unit/api/methods/test_client.py
